requirements.txt
torch>=1.8.0
torchvision>=0.9.0
diffusers>=0.2.4
datasets>=1.2.1
accelerate>=0.5.1
Pillow>=8.1.0
matplotlib>=3.3.4
tqdm>=4.59.0
numpy>=1.19.5
